requirements.txt
amqp==2.5.0
APScheduler==3.6.0
astroid==2.2.5
billiard==3.6.0.0
celery==4.3.0
Click==7.0
coverage==4.5.3
isort==4.3.20
kombu==4.6.0
lazy-object-proxy==1.4.1
mccabe==0.6.1
pkg-resources==0.0.0
pylint==2.3.1
python-dotenv==0.10.2
pytz==2019.1
six==1.12.0
typed-ast==1.3.5
tzlocal==1.5.1
vine==1.3.0
wrapt==1.11.1
